+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `Patch.patchAndAddFile(...)` utility.
+
 ### [v0.03](https://github.com/realityforge/gir/tree/v0.03) (2018-03-23)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.02...v0.03)